Add Workflow Comments (#1417)
Original PR #1293
-Updating Comments related stuff to the latest master configs and environment

## Introduction

This PR includes the basic features of the Workflow Commenting feature of Texera--a feature implemented to support communication between contributors of a workflow and comprehension of people reading a workflow.

Early Stage :
- Now can add `commentBox` objects onto the canvas
- a pop-up window will be displayed when double-clicking the comment icon
- delete button on the top right corner which allows you to delete the commentBox
- after opening the pop-up window, you can edit comments by typing in the input box.
- the input box stores newly added comments and previous comments in a list.
![Comments-Front-End](https://user-images.githubusercontent.com/72828385/156052516-c626c8c1-efb1-49c6-a2a1-9a9da3437f16.gif)

## Lifecycle of Comments
A comment (the little dialog icon on canvas) consists of two components: a **CommentBox** and **a list of Comments**. 
**CommentBox** is generated when the dialog icon pops up, it is a container object with information of an id as identifier and a point representing its position on the canvas, and an empty list of **Comments**.
**Comment** is generated and added to the list in **CommentBox** when someone types a comment and hits the send button. it contains information related to its creator, its content, and its creation time.
Refer to `workflow-common.interface.ts` for more details.
When the **Delete** button on the top-right corner of the icon is clicked, the commentBox gets deleted as well as all its corresponding comments.

## Storage of Comments
Currently, CommentBoxes are components of a workflow (parallel with Links and Operators), it is persisted into the backend database as an attribute in the JSON object with every information embedded in -- position, ID, and all the comments inside. @Yicong-Huang, @zuozhiw and I have been discussing whether we should seperate comments from commentBoxes since  the current method lacks the ability to handle detailed requests related to specific comments.

## 2022/02/26 Update
- Updated to work with the share-editing feature.
- Optimized front-end CommentBox UI for user-friendly purposes.

+++ core/new-gui/src/app/workspace/component/workflow-editor/workflow-editor.component.ts
@@ -4,6 +4,8 @@ import * as joint from "jointjs";
 // 2) always add this import statement even if TypeScript doesn't show an error https://github.com/Microsoft/TypeScript/issues/22016
 import * as jQuery from "jquery";
 import { fromEvent, merge } from "rxjs";
+import { NzModalCommentBoxComponent } from "./comment-box-modal/nz-modal-comment-box.component";
+import { NzModalRef, NzModalService } from "ng-zorro-antd/modal";
 import { assertType } from "src/app/common/util/assert";
 import { environment } from "../../../../environments/environment";
 import { DragDropService } from "../../service/drag-drop/drag-drop.service";
@@ -114,7 +116,8 @@ export class WorkflowEditorComponent implements AfterViewInit {
     private jointUIService: JointUIService,
     private workflowStatusService: WorkflowStatusService,
     private workflowUtilService: WorkflowUtilService,
-    private executeWorkflowService: ExecuteWorkflowService
+    private executeWorkflowService: ExecuteWorkflowService,
+    private nzModalService: NzModalService
   ) {}
 
   public getJointPaper(): joint.dia.Paper {
@@ -527,6 +530,7 @@ export class WorkflowEditorComponent implements AfterViewInit {
   }
 
   private handleCellHighlight(): void {
+    this.handleHighlightMouseDBClickInput();
     this.handleHighlightMouseInput();
     this.handleElementHightlightEvent();
   }
@@ -555,6 +559,23 @@ export class WorkflowEditorComponent implements AfterViewInit {
       });
   }
 
+  private handleHighlightMouseDBClickInput(): void {
+    fromEvent<JointPaperEvent>(this.getJointPaper(), "cell:pointerdblclick")
+      .pipe(untilDestroyed(this))
+      .subscribe(event => {
+        const clickedCommentBox = event[0].model;
+        if (
+          clickedCommentBox.isElement() &&
+          this.workflowActionService.getTexeraGraph().hasCommentBox(clickedCommentBox.id.toString())
+        ) {
+          this.workflowActionService.getJointGraphWrapper().setMultiSelectMode(<boolean>event[1].shiftKey);
+          const elementID = event[0].model.id.toString();
+          if (this.workflowActionService.getTexeraGraph().hasCommentBox(elementID)) {
+            this.workflowActionService.getJointGraphWrapper().highlightCommentBoxes(elementID);
+          }
+        }
+      });
+  }
   /**
    * Handles user mouse down events to trigger logically highlight and unhighlight an operator or group.
    * If user clicks the operator/group while pressing the shift key, multiselect mode is turned on.
@@ -657,6 +678,40 @@ export class WorkflowEditorComponent implements AfterViewInit {
           this.getJointPaper().findViewByModel(elementID).unhighlight("rect.body", { highlighter: highlightOptions })
         )
       );
+
+    this.workflowActionService
+      .getJointGraphWrapper()
+      .getJointCommentBoxHighlightStream()
+      .pipe(untilDestroyed(this))
+      .subscribe(commentBoxIDs => {
+        this.openCommentBox(commentBoxIDs[0]);
+      });
+  }
+
+  private openCommentBox(commentBoxID: string): void {
+    const commentBox = this.workflowActionService.getTexeraGraph().getCommentBox(commentBoxID);
+    const modalRef: NzModalRef = this.nzModalService.create({
+      // modal title
+      nzTitle: "Comments",
+      nzContent: NzModalCommentBoxComponent,
+      // set component @Input attributes
+      nzComponentParams: {
+        // set the index value and page size to the modal for navigation
+        commentBox: commentBox,
+      },
+      // prevent browser focusing close button (ugly square highlight)
+      nzAutofocus: null,
+      // modal footer buttons
+      nzFooter: [
+        {
+          label: "OK",
+          onClick: () => {
+            modalRef.destroy();
+          },
+          type: "primary",
+        },
+      ],
+    });
   }
 
   private handleOperatorSuggestionHighlightEvent(): void {
@@ -728,7 +783,12 @@ export class WorkflowEditorComponent implements AfterViewInit {
       )
       .pipe(untilDestroyed(this))
       .subscribe(elementView => {
-        this.workflowActionService.deleteOperator(elementView.model.id.toString());
+        if (this.workflowActionService.getTexeraGraph().hasOperator(elementView.model.id.toString())) {
+          this.workflowActionService.deleteOperator(elementView.model.id.toString());
+        }
+        if (this.workflowActionService.getTexeraGraph().hasCommentBox(elementView.model.id.toString())) {
+          this.workflowActionService.deleteCommentBox(elementView.model.id.toString());
+        }
       });
   }
